+++ Main/src/main/java/views/ExampleAppEquationView.java
@@ -3,7 +3,6 @@ import controllers.Controller;
 
 public class ExampleAppEquationView {
 
-
     public void startApp(String[] args) {
         Controller contr = new Controller(args);
         contr.getDecision();
